+++ CHANGES.md
@@ -1,6 +1,7 @@
 [unreleased]
 * update readme
 * sort listed plugins in 'Report a Bug'
+* add mu-plugins in 'Report a Bug'
 
 #### 3.3.0 / 2023-03-16
 * added `Report a Bug` feature, thanks @costdev, @ironprogrammer